Random/alg_trading/requirements.txt
pandas
matplotlib
numpy
scikit-learn
jupyter
beautifulsoup4
seaborn
boto3
ruamel.yaml
requests
pandas_datareader
pyarrow
psycopg2
